insurance-product-backend/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
pydantic==2.10.5
python-multipart==0.0.20
sqlalchemy==2.0.23
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
